Add support for io_uring with Netty 4.2.1.Final
Netty has finally added io_uring support as part of the 4.2 release. In this update, they've changed the EventLoop class hierarchy, deprecating Epoll/KQueue/Nio EventLoopGroup and replacing it with Multi/SingleThreadIoEventLoopGroup composed of IoHandlerFactory, which defines the event framework. No IOURingEventLoopGroup was ever added.

NettyEventLoops needs to adopt a new approach to deciding which EventLoopType we are using. Using isIoType(Handler) seems like a reasonable way to do this.

I've also removed usages of the deprecated EventLoopGroup classes and replaced them with the new way of doing things.